requirements.txt
accelerate
pandas
matplotlib
seaborn
tensorflow
scikit-learn
transformers
numpy
torch
Flask
Pillow
gunicorn
safetensors